+++ customer_apple_pay_integration_test.go
@@ -8,7 +8,7 @@ import (
 	"reflect"
 	"testing"
 
-	"github.com/braintree-go/braintree-go/testhelpers"
+	"github.com/Figure53/braintree-go/testhelpers"
 )
 
 func TestCustomerApplePayCard(t *testing.T) {